Move back tryInstallObject
