+++ src/OpenLoco/Objects/ObjectIndex.h
@@ -26,6 +26,6 @@ namespace OpenLoco::ObjectManager
     void loadIndex();
 
     std::vector<std::pair<uint32_t, ObjectIndexEntry>> getAvailableObjects(ObjectType type);
-    bool tryInstallObject(const ObjectHeader& object, stdx::span<const uint8_t> data);
+    bool isObjectInstalled(const ObjectHeader& objectHeader);
     ObjIndexPair getActiveObject(ObjectType objectType, uint8_t* edi);
 }
