+++ src/OpenLoco/Objects/ObjectManager.cpp
@@ -8,6 +8,8 @@
 #include "../Localisation/FormatArguments.hpp"
 #include "../Localisation/StringIds.h"
 #include "../S5/SawyerStream.h"
+#include "../Ui.h"
+#include "../Ui/ProgressBar.h"
 #include "../Utility/Numeric.hpp"
 #include "../Utility/Stream.hpp"
 #include "AirportObject.h"
@@ -176,7 +178,7 @@ namespace OpenLoco::ObjectManager
         drawPreview,
     };
 
-    bool callObjectValidate(const ObjectType type, Object& obj)
+    static bool callObjectValidate(const ObjectType type, Object& obj)
     {
         switch (type)
         {
@@ -503,6 +505,36 @@ namespace OpenLoco::ObjectManager
         throw std::runtime_error("Object not loaded at this index");
     }
 
+    // 0x00472754
+    static uint32_t computeChecksum(stdx::span<const uint8_t> data, uint32_t seed)
+    {
+        auto checksum = seed;
+        for (auto d : data)
+        {
+            checksum = Utility::rol(checksum ^ d, 11);
+        }
+        return checksum;
+    }
+
+    // 0x0047270B
+    static bool computeObjectChecksum(const ObjectHeader& object, stdx::span<const uint8_t> data)
+    {
+        // Compute the checksum of header and data
+
+        // Annoyingly the header you need to only compute the first byte of the flags
+        stdx::span<const uint8_t> headerFlag(reinterpret_cast<const uint8_t*>(&object), 1);
+        auto checksum = computeChecksum(headerFlag, objectChecksumMagic);
+
+        // And then the name
+        stdx::span<const uint8_t> headerName(reinterpret_cast<const uint8_t*>(&object.name), sizeof(ObjectHeader::name));
+        checksum = computeChecksum(headerName, checksum);
+
+        // Finally compute the datas checksum
+        checksum = computeChecksum(data, checksum);
+
+        return checksum == object.checksum;
+    }
+
     // 0x00471BC5
     static bool load(const ObjectHeader& header, LoadedObjectId id)
     {
@@ -651,37 +683,7 @@ namespace OpenLoco::ObjectManager
         return result;
     }
 
-    // 0x00472754
-    static uint32_t computeChecksum(stdx::span<const uint8_t> data, uint32_t seed)
-    {
-        auto checksum = seed;
-        for (auto d : data)
-        {
-            checksum = Utility::rol(checksum ^ d, 11);
-        }
-        return checksum;
-    }
-
-    // 0x0047270B
-    bool computeObjectChecksum(const ObjectHeader& object, stdx::span<const uint8_t> data)
-    {
-        // Compute the checksum of header and data
-
-        // Annoyingly the header you need to only compute the first byte of the flags
-        stdx::span<const uint8_t> headerFlag(reinterpret_cast<const uint8_t*>(&object), 1);
-        auto checksum = computeChecksum(headerFlag, objectChecksumMagic);
-
-        // And then the name
-        stdx::span<const uint8_t> headerName(reinterpret_cast<const uint8_t*>(&object.name), sizeof(ObjectHeader::name));
-        checksum = computeChecksum(headerName, checksum);
-
-        // Finally compute the datas checksum
-        checksum = computeChecksum(data, checksum);
-
-        return checksum == object.checksum;
-    }
-
-    bool partialLoad(const ObjectHeader& header, stdx::span<uint8_t> objectData)
+    static bool partialLoad(const ObjectHeader& header, stdx::span<uint8_t> objectData)
     {
         auto type = header.getType();
         size_t index = 0;
@@ -702,6 +704,146 @@ namespace OpenLoco::ObjectManager
         getRepositoryItem(type).objects[index] = obj;
         getRepositoryItem(type).object_entry_extendeds[index] = ObjectEntry2(header, objectData.size());
         return true;
+    }
+
+    static void permutateObjectFilename(std::string& filename)
+    {
+        auto* firstChar = filename.c_str();
+        auto* endChar = &filename[filename.size()];
+        auto* c = endChar;
+        do
+        {
+            c--;
+            if (c == firstChar)
+            {
+                filename = "00000000";
+                break;
+            }
+            if (*c < '0')
+            {
+                *c = '/';
+            }
+            if (*c == '9')
+            {
+                *c = '@';
+            }
+            if (*c == 'Z')
+            {
+                *c = '/';
+            }
+            *c = *c + 1;
+        } while (*c == '0');
+    }
+
+    // All object files are based on their internal object header name but
+    // there is a chance of a name collision this function works out if the name
+    // is possible and if not permutates the name until it is valid.
+    static fs::path findObjectPath(std::string& filename)
+    {
+        auto objPath = Environment::getPath(Environment::PathId::objects);
+
+        bool permutateName = false;
+        do
+        {
+            if (permutateName)
+            {
+                permutateObjectFilename(filename);
+            }
+            objPath.replace_filename(filename);
+            objPath.replace_extension(".DAT");
+            permutateName = true;
+        } while (fs::exists(objPath));
+        return objPath;
+    }
+
+    static void sanatiseObjectFilename(std::string& filename)
+    {
+        // Trim string at first space (note this copies vanilla but maybe shouldn't)
+        auto space = filename.find_first_of(' ');
+        if (space != std::string::npos)
+        {
+            filename = filename.substr(0, space);
+        }
+        // Make filename uppercase
+        std::transform(std::begin(filename), std::end(filename), std::begin(filename), toupper);
+    }
+
+    // 0x0047285C
+    static bool installObject(const ObjectHeader& objectHeader)
+    {
+        // Prepare progress bar
+        char caption[512];
+        auto* str = StringManager::formatString(caption, sizeof(caption), StringIds::installing_new_data);
+        // Convert object name to string so it is properly terminated
+        std::string objectname(objectHeader.getName());
+        strcat(str, objectname.c_str());
+        Ui::ProgressBar::begin(caption);
+        Ui::ProgressBar::setProgress(50);
+        Ui::processMessagesMini();
+
+        // Get new file path
+        std::string filename = objectname;
+        sanatiseObjectFilename(filename);
+        auto objPath = findObjectPath(filename);
+
+        // Create new file and output object file
+        Ui::ProgressBar::setProgress(180);
+        SawyerStreamWriter stream(objPath);
+        writePackedObjects(stream, { objectHeader });
+
+        // Free file
+        stream.close();
+        Ui::ProgressBar::setProgress(240);
+        Ui::ProgressBar::setProgress(255);
+        Ui::ProgressBar::end();
+        return true;
+    }
+
+    // 0x00472687 based on
+    bool tryInstallObject(const ObjectHeader& objectHeader, stdx::span<const uint8_t> data)
+    {
+        unloadAll();
+        if (!computeObjectChecksum(objectHeader, data))
+        {
+            return false;
+        }
+        // Copy the object into Loco freeable memory (required for when partialLoad loads the object)
+        uint8_t* objectData = static_cast<uint8_t*>(malloc(data.size()));
+        if (objectData == nullptr)
+        {
+            return false;
+        }
+        std::copy(std::begin(data), std::end(data), objectData);
+
+        auto* obj = reinterpret_cast<Object*>(objectData);
+        if (!callObjectValidate(objectHeader.getType(), *obj))
+        {
+            return false;
+        }
+
+        if (_totalNumImages >= Gfx::G1ExpectedCount::kObjects + Gfx::G1ExpectedCount::kDisc)
+        {
+            // Free objectData?
+            return false;
+        }
+
+        // Warning this saves a copy of the objectData pointer and must be unloaded prior to exiting this function
+        if (!partialLoad(objectHeader, stdx::span(objectData, data.size())))
+        {
+            return false;
+        }
+
+        // Object already installed so no need to install it
+        if (isObjectInstalled(objectHeader))
+        {
+            unloadAll();
+            return false;
+        }
+
+        bool result = installObject(objectHeader);
+
+        unloadAll();
+        return result;
     }
 
     static constexpr SawyerEncoding getBestEncodingForObjectType(ObjectType type)
